chore: moved noView and inlineView to compat package
