+++ packages/aurelia/src/index.ts
@@ -384,8 +384,6 @@ export {
   // IElementProjector,
   // IProjectorLocator,
   useShadowDOM,
-  noView,
-  inlineView,
 
   // ISanitizer,
   // SanitizeValueConverter,
