+++ packages/runtime-html/src/index.ts
@@ -244,8 +244,6 @@ export {
   type CustomElementStaticAuDefinition,
   useShadowDOM,
   processContent,
-  noView,
-  inlineView,
 } from './resources/custom-element';
 
 export {
